- Use int instead of byte to facilitate use - Specify formatter class name as string or null if no formatter   is used

+++ NetworkProtocolModeler/src/runtime/Serializer.java
@@ -4,13 +4,13 @@ import java.lang.reflect.Field;
 
 public class Serializer {
 
-	public static void serializeToBytes(long val, byte bytes, byte[] dest,
+	public static void serializeToBytes(long val, int bytes, byte[] dest,
 			byte off) {
 		for (int i = 0; i < bytes; i++)
 			dest[off + i] = (byte) ((val >> (i * 8)) & 0b11111111);
 	}
 
-	public static void serializeToBytes(String val, byte bytes, byte[] dest,
+	public static void serializeToBytes(String val, int bytes, byte[] dest,
 			byte off) {
 		byte[] src = val.getBytes();
 		if (bytes == 0)
@@ -19,7 +19,7 @@ public class Serializer {
 			dest[off + i] = src[i];
 	}
 
-	public static void serializeToBytes(byte[] val, byte bytes, byte[] dest,
+	public static void serializeToBytes(byte[] val, int bytes, byte[] dest,
 			byte off) {
 		if (bytes == 0)
 			bytes = (byte) val.length;
